requirements.txt
discord.py
asyncio
boto